requirements.txt
pdf417==0.8.1
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0